Move plan nodes to nodes subdirectory
Reorg source code to create a nodes subdirectory.
Move plan nodes to this drectory.
