+++ src/copy.c
@@ -38,8 +38,8 @@
 #include "hypertable.h"
 #include "copy.h"
 #include "dimension.h"
-#include "chunk_insert_state.h"
-#include "chunk_dispatch.h"
+#include "nodes/chunk_insert_state.h"
+#include "nodes/chunk_dispatch.h"
 #include "subspace_store.h"
 #include "compat.h"
 #include "cross_module_fn.h"
