+++ src/nodes/constraint_aware_append/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(SOURCES
+  ${CMAKE_CURRENT_SOURCE_DIR}/constraint_aware_append.c
+)
+target_sources(${PROJECT_NAME} PRIVATE ${SOURCES})
