+++ src/planner.c
@@ -52,9 +52,9 @@
 #include "utils.h"
 #include "guc.h"
 #include "dimension.h"
-#include "chunk_dispatch_plan.h"
-#include "hypertable_insert.h"
-#include "constraint_aware_append.h"
+#include "nodes/chunk_dispatch_plan.h"
+#include "nodes/hypertable_insert.h"
+#include "nodes/constraint_aware_append/constraint_aware_append.h"
 #include "chunk_append/chunk_append.h"
 #include "partitioning.h"
 #include "dimension_slice.h"
