+++ tsl/src/nodes/skip_scan/planner.c
@@ -28,7 +28,7 @@
 #include <import/planner.h>
 #include "guc.h"
 #include "nodes/skip_scan/skip_scan.h"
-#include "constraint_aware_append.h"
+#include "nodes/constraint_aware_append/constraint_aware_append.h"
 #include "chunk_append/chunk_append.h"
 
 #include <math.h>
